mgr_util: add CephfsClient implementation
This pulls parts of the VolumesClient implementation into mgr_util to
make the CephFS specific pieces available to other mgr modules. To
reduce code duplication the VolumeClient now extends the CephfsClient
class to add the volume specific methods.

(cherry picked from commit a44de38b61d598fb0512ea48da0de4179d39b804)
(cherry picked from commit f87a611e74776cca0969b96ce135499369a40679)

Conflicts:
	src/pybind/mgr/volumes/fs/volume.py
